Make Relations higher priority in parsing (#5)
Due to the short notation of some UMLElements, e.g. components `[component]`, and the power of the `EndLine` verb. Relations also declaring component `[componentA] --> [componentB]` were detected as just the `[componentA]`. Maybe the right approach here would be to not use `EndLine` as often as we do right now, but this might have an impact on the robustness against non implemented syntax features.